chore: implement JujubaSVG composable and refactor SVG handling in RootApp
